move eventDecoder to root folder && fix devent decoding
